Refactor to build and use Codec classes
