+++ pbj-core/pbj-compiler/src/main/java/com/hedera/pbj/compiler/impl/LookupHelper.java
@@ -160,8 +160,7 @@ public final class LookupHelper {
 			return  switch (fileType) {
 				case MODEL, PROTOC -> name;
 				case SCHEMA -> name + FileAndPackageNamesConfig.SCHEMA_JAVA_FILE_SUFFIX;
-				case PARSER -> name + FileAndPackageNamesConfig.PARSER_JAVA_FILE_SUFFIX;
-				case WRITER -> name + FileAndPackageNamesConfig.WRITER_JAVA_FILE_SUFFIX;
+				case CODEC -> name + FileAndPackageNamesConfig.CODEC_JAVA_FILE_SUFFIX;
 				case TEST -> name + FileAndPackageNamesConfig.TEST_JAVA_FILE_SUFFIX;
 			};
 		}
@@ -195,8 +194,7 @@ public final class LookupHelper {
 					//noinspection ConstantConditions
 					case MODEL, PROTOC -> basePackage;
 					case SCHEMA -> basePackage + '.' + FileAndPackageNamesConfig.SCHEMAS_SUBPACKAGE;
-					case PARSER -> basePackage + '.' + FileAndPackageNamesConfig.PARSERS_SUBPACKAGE;
-					case WRITER -> basePackage + '.' + FileAndPackageNamesConfig.WRITERS_SUBPACKAGE;
+					case CODEC -> basePackage + '.' + FileAndPackageNamesConfig.CODECS_SUBPACKAGE;
 					case TEST -> basePackage + '.' + FileAndPackageNamesConfig.TESTS_SUBPACKAGE;
 				};
 			}
