+++ pbj-integration-tests/src/main/java/com/hedera/pbj/integration/EverythingWriterPerfTest.java
@@ -1,8 +1,8 @@
 package com.hedera.pbj.integration;
 
 import com.hedera.pbj.runtime.io.DataBuffer;
+import com.hedera.pbj.test.proto.pbj.Everything;
 import com.hederahashgraph.api.proto.java.GetAccountDetailsResponse;
-import com.hederahashgraph.api.proto.pbj.test.writer.EverythingWriter;
 
 @SuppressWarnings("DuplicatedCode")
 public class EverythingWriterPerfTest {
@@ -12,7 +12,7 @@ public class EverythingWriterPerfTest {
 
         for (int i = 0; i < 10_000_000; i++) {
             outDataBuffer.reset();
-            EverythingWriter.write(EverythingTestData.EVERYTHING, outDataBuffer);
+            Everything.PROTOBUF.write(EverythingTestData.EVERYTHING, outDataBuffer);
             if (outDataBuffer.getPosition() <= 0) {
                 System.out.println("outDataBuffer = " + outDataBuffer);
             }
@@ -21,7 +21,7 @@ public class EverythingWriterPerfTest {
     public static void main2(String[] args) throws Exception {
         // write to temp data buffer and then read into byte array
         DataBuffer tempDataBuffer = DataBuffer.allocate(5 * 1024 * 1024, false);
-        EverythingWriter.write(EverythingTestData.EVERYTHING, tempDataBuffer);
+        Everything.PROTOBUF.write(EverythingTestData.EVERYTHING, tempDataBuffer);
         tempDataBuffer.flip();
         final byte[] protobuf = new byte[(int) tempDataBuffer.getRemaining()];
         tempDataBuffer.readBytes(protobuf);
