+++ pbj-integration-tests/src/main/proto/everything.proto
@@ -2,9 +2,9 @@ syntax = "proto3";
 
 package proto;
 
-option java_package = "com.hederahashgraph.api.proto.java.test";
+option java_package = "com.hedera.pbj.test.proto.java";
 option java_multiple_files = true;
-// <<<pbj.java_package = "com.hederahashgraph.api.proto.pbj.test">>> This comment is special code for setting PBJ Compiler java package
+// <<<pbj.java_package = "com.hedera.pbj.test.proto.pbj">>> This comment is special code for setting PBJ Compiler java package
 
 import "timestampTest.proto";
 import "google/protobuf/wrappers.proto";
