+++ pbj-integration-tests/src/main/proto/extendedUtf8StingTest.proto
@@ -22,9 +22,9 @@ package proto;
  * ‍
  */
 
-option java_package = "com.hederahashgraph.api.proto.java.test";
+option java_package = "com.hedera.pbj.test.proto.java";
 option java_multiple_files = true;
-// <<<pbj.java_package = "com.hederahashgraph.api.proto.pbj.test">>> This comment is special code for setting PBJ Compiler java package
+// <<<pbj.java_package = "com.hedera.pbj.test.proto.pbj">>> This comment is special code for setting PBJ Compiler java package
 
 /**
  * Simple message with a string for extended UTF8 testing
